Convert standalone controllers in LibraryFooter and ExplorationGraph to component directives (#6840)
* Exploration Graph

* Library footer

* Oppia library footer

* Fix lint

* Fix CODEOWNERS

* Added missing EOL
