+++ core/templates/dev/head/pages/about/about.html
@@ -632,5 +632,5 @@
 {% endblock footer_js %}
 
 {% block footer%}
-  {% include 'pages/footer.html' %}
+  <oppia-footer></oppia-footer>
 {% endblock %}
