+++ core/templates/dev/head/pages/contact/Contact.ts
@@ -17,3 +17,4 @@
  */
 
 require('components/background/BackgroundBannerDirective.ts');
+require('pages/OppiaFooterDirective.ts');
