+++ core/templates/dev/head/pages/contact/contact.html
@@ -106,5 +106,5 @@
 {% endblock footer_js %}
 
 {% block footer%}
-  {% include 'pages/footer.html' %}
+  <oppia-footer></oppia-footer>
 {% endblock %}
