+++ core/templates/dev/head/pages/donate/donate.html
@@ -164,5 +164,5 @@
 {% endblock footer_js %}
 
 {% block footer %}
-  {% include 'pages/footer.html' %}
+  <oppia-footer></oppia-footer>
 {% endblock %}
