+++ core/templates/dev/head/pages/error/Error.ts
@@ -16,6 +16,8 @@
  * @fileoverview Controllers for the error page.
  */
 
+require('pages/OppiaFooterDirective.ts');
+
 require('domain/utilities/UrlInterpolationService.ts');
 require('services/PageTitleService.ts');
