+++ core/templates/dev/head/pages/error/error.html
@@ -72,5 +72,5 @@
 {% endblock footer_js %}
 
 {% block footer %}
-  {% include 'pages/footer.html' %}
+  <oppia-footer></oppia-footer>
 {% endblock %}
