+++ core/templates/dev/head/pages/exploration_editor/editor_tab/editor_tab.html
@@ -4,7 +4,7 @@
   </div>
   <div class="col-lg-4 col-md-4 col-sm-4">
     <div class="oppia-editor-sidebar hidden-xs hidden-sm">
-      {% include 'pages/exploration_editor/editor_tab/exploration_graph.html' %}
+      <exploration-graph></exploration-graph>
       <unresolved-answers-overview></unresolved-answers-overview>
     </div>
   </div>
