--- core/templates/dev/head/pages/exploration_editor/editor_tab/exploration_graph.html
@@ -1,12 +1,0 @@
-<div ng-controller="ExplorationGraph" ng-show="isGraphShown()" class="oppia-state-graph-animate-show">
-  <span style="font-size: 16px;">
-    <strong>Exploration Overview</strong>
-  </span>
-
-  <md-card style="background-color: white; margin: 20px 0px; padding: 8px;">
-    <div class="oppia-state-graph-container protractor-test-exploration-graph">
-      <div state-graph-visualization graph-data="getGraphData()" current-state-id="getActiveStateName()" on-click-function="onClickStateInMinimap" on-delete-function="deleteState" on-maximize-function="openStateGraphModal" allow-panning="true" center-at-current-state="true" style="height: 100%;" is-editable="isEditable()" show-warning-sign="true">
-      </div>
-    </div>
-  </md-card>
-</div>
