+++ core/templates/dev/head/pages/exploration_editor/editor_tab/exploration_graph_directive.html
@@ -1,0 +1,12 @@
+<div ng-show="$ctrl.isGraphShown()" class="oppia-state-graph-animate-show">
+  <span style="font-size: 16px;">
+    <strong>Exploration Overview</strong>
+  </span>
+
+  <md-card style="background-color: white; margin: 20px 0px; padding: 8px;">
+    <div class="oppia-state-graph-container protractor-test-exploration-graph">
+      <div state-graph-visualization graph-data="$ctrl.getGraphData()" current-state-id="$ctrl.getActiveStateName()" on-click-function="$ctrl.onClickStateInMinimap" on-delete-function="$ctrl.deleteState" on-maximize-function="$ctrl.openStateGraphModal" allow-panning="true" center-at-current-state="true" style="height: 100%;" is-editable="$ctrl.isEditable()" show-warning-sign="true">
+      </div>
+    </div>
+  </md-card>
+</div>
