+++ core/templates/dev/head/pages/get_started/get_started.html
@@ -98,5 +98,5 @@
 {% endblock footer_js %}
 
 {% block footer%}
-  {% include 'pages/footer.html' %}
+  <oppia-footer></oppia-footer>
 {% endblock %}
