+++ core/templates/dev/head/pages/library/library.html
@@ -25,11 +25,7 @@
 {% endblock %}
 
 {% block footer %}
-  <div ng-controller="LibraryFooter">
-    <div ng-if="footerIsDisplayed">
-      {% include 'pages/footer.html' %}
-    </div>
-  </div>
+  <library-footer></library-footer>
 {% endblock footer %}
 
 {% block footer_js %}
