+++ core/templates/dev/head/pages/library/library_footer_directive.html
@@ -1,0 +1,5 @@
+<div>
+  <div ng-if="$ctrl.footerIsDisplayed">
+    <oppia-footer></oppia-footer>
+  </div>
+</div>
