+++ core/templates/dev/head/pages/oppia_footer_directive.html
