+++ core/templates/dev/head/pages/preferences/Preferences.ts
@@ -20,6 +20,7 @@ require('components/forms/Select2DropdownDirective.ts');
 require('components/forms/ImageUploaderDirective.ts');
 require('components/background/BackgroundBannerDirective.ts');
 require('filters/TruncateFilter.ts');
+require('pages/OppiaFooterDirective.ts');
 
 require('domain/utilities/LanguageUtilService.ts');
 require('domain/utilities/UrlInterpolationService.ts');
