+++ core/templates/dev/head/pages/preferences/preferences.html
@@ -272,7 +272,7 @@
 {% endblock %}
 
 {% block footer %}
-  {% include 'pages/footer.html' %}
+  <oppia-footer></oppia-footer>
 {% endblock %}
 
 {% block footer_js %}
