+++ core/templates/dev/head/pages/privacy/privacy.html
@@ -432,5 +432,5 @@
 {% endblock %}
 
 {% block footer %}
-  {% include 'pages/footer.html' %}
+  {% include 'pages/oppia_footer_directive.html' %}
 {% endblock %}
