+++ core/templates/dev/head/pages/profile/profile.html
@@ -423,7 +423,7 @@
 {% endblock %}
 
 {% block footer %}
-  {% include 'pages/footer.html' %}
+  <oppia-footer></oppia-footer>
 {% endblock %}
 
 {% block footer_js %}
