+++ core/templates/dev/head/pages/splash/splash.html
@@ -198,7 +198,7 @@
 {% endblock %}
 
 {% block footer %}
-  {% include 'pages/footer.html' %}
+  <oppia-footer></oppia-footer>
 {% endblock %}
 
 {% block footer_js %}
