+++ core/templates/dev/head/pages/splash/splash_at1.html
@@ -205,7 +205,7 @@
 {% endblock %}
 
 {% block footer %}
-  {% include 'pages/footer.html' %}
+  <oppia-footer></oppia-footer>
 {% endblock %}
 
 {% block footer_js %}
