+++ core/templates/dev/head/pages/teach/teach.html
@@ -162,5 +162,5 @@
 {% endblock footer_js %}
 
 {% block footer%}
-  {% include 'pages/footer.html' %}
+  <oppia-footer></oppia-footer>
 {% endblock %}
