+++ core/templates/dev/head/pages/terms/terms.html
@@ -320,5 +320,5 @@
 {% endblock %}
 
 {% block footer%}
-  {% include 'pages/footer.html' %}
+  {% include 'pages/oppia_footer_directive.html' %}
 {% endblock %}
