+++ core/templates/dev/head/pages/thanks/thanks.html
@@ -57,5 +57,5 @@
 {% endblock footer_js %}
 
 {% block footer%}
-  {% include 'pages/footer.html' %}
+  <oppia-footer></oppia-footer>
 {% endblock %}
